Add files via upload
HOW TO USE >> To use the application, first create a directory in C drive called "VICINITY".
Copy contents of zip package to c:/VICINITY and you're ready to run the app.

Run the run.exe file in both the sender and receiver computer, make sure that both computers are connected to the same wifi or hotspot.
Click "Receive" on the receiving computer.. on the sending computer, copy files to send folder and then click on send.

The app will dynamically connect the two computers and then open the camera terminal in both the computers. Do a "Grab" gesture on 
the sender computer and with a "Release" gesture, drop the files into the receiver pc.